% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AGFCEM}
\alias{AGFCEM}
\title{Dummy}
\usage{
AGFCEM(input)
}
\arguments{
\item{input}{a numerical input}
}
\value{
Nothing.
}
\description{
This is a dummy function.
}
\details{
This is a dummy function.
}
\examples{
AGFCEM(3.0)
}
