% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleData-class.R
\docType{class}
\name{EnsembleData-class}
\alias{EnsembleData-class}
\title{A class to hold the Ensemble data}
\description{
A class that holds the observation data, simulator outputs, and prior information
ready to fit the ensemble model in Stan.
}
\details{
A new \code{EnsembleData} object can be created with the \code{\link[=EnsembleData]{EnsembleData()}}
constructor, but it is not necessary to do this if using the \code{fit_ensemble_model}
function, as it will be created automatically.

There is only one slot for this object, containing a list of data in the required
form to fit the Stan model.
}
\section{Slots}{

\describe{
\item{\code{stan_input}}{A list of parameters in the correct form to fit the ensemble
model in Stan.}

\item{\code{observations}}{A \code{list} of observations and associated covariances.
See the \code{EnsembleData} constructor details.}

\item{\code{simulators}}{A \code{list} of simulator output and covariance pairs. See the \code{EnsembleData} constructor details.}

\item{\code{priors}}{A \code{list} of data specifying the prior distributions for the ensemble.
See the \code{EnsembleData} constructor details.}
}}

