% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleData-class.R
\name{EnsembleData}
\alias{EnsembleData}
\title{Constructor for the \code{EnsembleData} class}
\usage{
EnsembleData(observations, simulators, priors)
}
\arguments{
\item{observations}{A \code{list} of observations and associated covariances.
See details.}

\item{simulators}{A \code{list} of simulator output and covariance pairs. See details.}

\item{priors}{A \code{list} of data specifying the prior distributions for the ensemble.
See the \code{define_priors} function for the required form.}
}
\value{
An object of type \linkS4class{EnsembleData}
}
\description{
A constructor for the \code{EnsembleData} class. This is used to convert
input data into the required form to fit the ensemble model using Stan.
}
\details{
Observation and covariance pairs should be passed through as a \code{list},
with observations / model outputs as the first element, and the covariance matrix
as the second. The observations / model outputs should be a \code{data.frame} (or a \code{matrix}) with each
column giving a different species and each row a different time. Rows should be named with
the times and columns should be named the appropriate variables. It is fine to have missing
years or species, however for any model output there should be at least one observation in the observation
data frame. Missing species / years are encoded by simply not including the corresponding row / column in the
data frame / matrix.
}
\examples{
N_species <- 4
priors <- define_priors(ind_st_var_params = list(25, 0.25),
                       ind_st_cor_form = "lkj", #Using an LKJ distribution for individual short-term discrepancies
                       ind_st_cor_params = 30, #The parameter is 30
                     ind_lt_var_params = list(rep(25,N_species),rep(0.25,N_species)),
                     ind_lt_cor_form = "beta",
                     ind_lt_cor_params = list(matrix(40,N_species, N_species), matrix(40, N_species, N_species)),
                     sha_st_var_exp = 3,
                     sha_st_cor_form = "lkj",
                     sha_st_cor_params = 30,
                   sha_lt_sd = rep(4,N_species))
ensemble_data <- EnsembleData(observations = list(SSB_obs, Sigma_obs),
                             simulators = list(list(SSB_ewe, Sigma_ewe, "EwE"),
                                           list(SSB_fs,  Sigma_fs, "FishSUMS"),
                                           list(SSB_lm,  Sigma_lm, "LeMans"),
                                           list(SSB_miz, Sigma_miz, "mizer")),
                              priors = priors)
}
