% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFit-class.R
\name{EnsembleFit}
\alias{EnsembleFit}
\title{Constructor for the \code{EnsembleFit} class}
\usage{
EnsembleFit(ensemble_data, samples = NULL, point_estimate = NULL)
}
\arguments{
\item{ensemble_data}{An \code{EnsembleData} object encapsulating the data used to fit the ensemble model.}

\item{samples}{A \code{stanfit} object containing the samples drawn from the fitted model. The default value is \code{NULL}.}

\item{point_estimate}{A \code{list} output of the optimised model. The default value is \code{NULL}.}
}
\value{
An object of type \code{EnsembleFit}
}
\description{
A constructor for the \code{EnsembleFit} class. This contains the samples or point estimates generated by the fitted model.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
\code{\linkS4class{EnsembleFit}}, \code{\link{fit_ensemble_model}}, \code{\link{generate_sample}}
}
