% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsemblePrior-class.R
\name{EnsemblePrior}
\alias{EnsemblePrior}
\title{Constructor for the \code{EnsemblePrior} class}
\usage{
EnsemblePrior(
  d,
  ind_st_params,
  ind_lt_params,
  sha_st_params,
  sha_lt_params,
  truth_params = list(10, list(10, 1), list(d, diag(d)))
)
}
\arguments{
\item{d}{A \code{numeric} specifying the number of species in the ensemble.}

\item{ind_st_params}{A \code{list} containing a prior specification for the individual short-term discrepancies \eqn{z_k^{(t)}}. See details}

\item{ind_lt_params}{A \code{list} containing a prior specification for the individual long-term discrepancies \eqn{\gamma_k}. See details}

\item{sha_st_params}{A \code{list} containing a prior specification for the shared short-term discrepancies \eqn{\eta^{(t)}}. See details}

\item{sha_lt_params}{A \code{numeric} containing the standard deviations for the normal prior used on the shared short-term discrepancy \eqn{\mu}. If a single value is supplied,  this is repeated for each species.}

\item{truth_params}{A \code{list} containing a prior specification for the processes on the truth \eqn{y^{(t)}}. The default value is \verb{list(10, list(10, 1), list(d, diag(d))} where \code{d} is the number of species.}
}
\value{
An object of class \code{EnsemblePrior}
}
\description{
A constructor for the \code{EnsemblePrior} class. This is used to encode prior information for the ensemble model.
}
\details{
Most discrepancy prior paramaters (\verb{ind_st_params, ind_lt_params, sha_st_params} but not \code{sha_lt_params}) should be encoded by a \code{list.} The entries of the list are respectively
\enumerate{
\item A \code{character} specifying how the priors are encoded. Currently supported priors are \code{'LKJ'}, \code{'inv_wishart'}, \code{'Beta'}, or \code{'hierarchical'} (\code{'hierarchical'} is only supported for the individual short-term discrepancies).
\item The prior parameters for the discrepancy variances, either a \code{list} of length 2 or a \code{numeric} of length 4. See below.
\item The correlation matrix parameters, either a \code{list} of length 2 or a \code{numeric} of length 4. See below.
}

There are currently four supported prior distributions on covariance matrices. As in Spence et. al. (2018), the discrepancy covariance matrices (individual and shared short-term discrepancies \eqn{z_k^{(t)},\eta^{(t)}} as well as individual long-term discrepancies \eqn{\gamma_k}, but not the shared long-term discrepancy \eqn{\delta}) are decomposed into a vector of variances and a correlation matrix \deqn{\Lambda = \sqrt{\mathrm{diag}(\pi)}  P \sqrt{\mathrm{diag}(\pi)},} where \eqn{\pi} is the vector of variances for each species, and \eqn{P} is the correlation matrix.

Selecting \code{'lkj'}, \code{'inv_wishart'}, \code{'beta'}, or \code{'hierarchical'} refers to setting LKJ, inverse Wishart, Beta, or hierarchical priors respectively. The variance parameters should be passed through as the second element of the \code{list} and the correlation parameters should be passed through as the third element. For \code{'lkj'}, \code{'inv_wishart'}, and \code{'beta'} selections, variances are parameterised by inverse-gamma distributions, so the second element of the \code{list} should itself be a \code{list} of length two, where each element gives the shape and scale parameters for each species (either as a single value which is the same for each species or a \code{numeric} with the same length as the number of species). The correlations should be in the following form:
\itemize{
\item If \code{'lkj'} is selected, then the third element should be a \code{numeric} \eqn{\eta} giving the LKJ shape parameter, such  that the probability density is given by  (Lewandowski et. al. 2009) \deqn{f(\Sigma | \eta)\propto \mathrm{det} (\Sigma)^{\eta - 1}.} Variances are parameterised by inverse-gamma distributions.
\item If \code{'inv_wishart'} is selected, then  the third element should be a \code{list} containing a scalar value \eqn{\nu} (giving the degrees of freedom) and a symmetric, positive definite matrix \eqn{\Sigma} (giving the scale matrix). The dimensions of \eqn{\Sigma} should be the same as the correlation matrix it produces (i.e \eqn{d \times d} where \eqn{d} is the number of species). The density of an inverse Wishart is given by  \deqn{f(W|\eta, S) = \frac{1}{2^{\eta d/2} \Gamma_N \left( \frac{\eta}{2} \right)} |S|^{\eta/2} |W|^{-(\eta + d + 1)/2}  \exp \left(- \frac{1}{2} \mathrm{tr}\left(SW^{-1} \right) \right),} where \eqn{\Gamma_N} is the multivariate gamma function and \eqn{\mathrm{tr \left(X \right)}} is the trace of \eqn{X}.  Note that inverse Wishart distributions act over the space of all covariance matrices. When used for a correlation  matrix, only the subset of valid covariance matrices that are also valid correlation matrices are considered. Variances are parameterised by inverse-gamma distributions.
\item If \code{'beta'} is selected, then  the third element should be a  \code{list} containing two \eqn{d \times d} matrices (where \eqn{d} is the number of species), giving the prior success parameters \eqn{\alpha} and prior failure parameters \eqn{\beta} respectively. To ensure positive-definiteness, the correlations are rescaled from \eqn{[-1,1] \rightarrow [0,1]} via the function \eqn{\frac{1}{\pi} \tan^{-1} \frac{\rho}{\sqrt{1-\rho^2}} + \frac{1}{2}}. It is on these rescaled parameters that the Beta distribution applies. Variances are parameterised by inverse-gamma distributions.
\item If \code{'hierarchical'} is selected, then variances are parameterised by hierarchical gamma distributions:
\deqn{\pi_{k, i} \sim \mathrm{Gamma}(a_{k, i}, b_{k, i})} where
\deqn{a_{k, i} \sim \mathrm{Gamma}(A, B)}
\deqn{b_{k, i} \sim \mathrm{Gamma}(C, D)}
The second element of the parameters \code{list} should then be a \code{numeric} of length 4, giving the \eqn{A, B, C, D} hyperparameters respectively. Correlations (\eqn{\rho_{k, i, j}} where \eqn{\rho_{k, i, j}} is the correlation between species \eqn{i} and \eqn{j} for the \eqn{k}th simulator) are parameterised by hierarchical Beta distributions.
\deqn{\frac{\rho_{k, i, j} + 1}{2} \sim \mathrm{Beta}(c_{k, i, j}, d_{k, i, j})} where
\deqn{c_{k, i, j} \sim \mathrm{Gamma}(\alpha, \beta)}
\deqn{d_{k, i, j} \sim \mathrm{Gamma}(\gamma, \delta)}
So the third element should be a \code{numeric} of length 4 giving the \eqn{\alpha, \beta, \gamma, \delta} hyperparameters. respectively. NOTE: This options is only supported for the individual short-term discrepancy terms.
}

In addition to priors on the discrepancy terms, it is also possible to add prior information on the truth. We require priors on the truth at \eqn{t=0}. By default, a \eqn{N(0, 10)} prior is used on the initial values, and an Inv-Gamma\eqn{(10, 1)} prior is used for the initial variances, however these values can be configured by the \code{truth_params} argument. The covariance matrix of the random walk of the truth \eqn{\Lambda_y} can also be configured subject to an inverse-Wishart prior. The elements of the \code{truth_params} list should be
\enumerate{
\item A \code{numeric} giving the standard deviation of the normal prior used for each species of the initial truth values. Default value is 10
\item A \code{list} of length \code{2} containing the shape and scale parameters (respectively) for the inverse gamma priors on the initial variance of the truth. The default value is \code{list(10, 1)}.
\item A \code{list} of length \code{2} containing the inverse-Wishart parameters for the random walk of the truth. The default value is \code{list(d, diag(d))} where \code{d} is the number of species.
}
}
\examples{
#Defining priors for a model with 4 species.
num_species <- 4
priors <- EnsemblePrior(
    d = num_species,
    ind_st_params = list("lkj",  list(3, 2), 3),
    ind_lt_params = list(
       "beta",
       list(c(10,4,8, 7),c(2,3,1, 4)),
       list(matrix(5, num_species, num_species),
            matrix(0.5, num_species, num_species))
     ),
    sha_st_params = list("inv_wishart",list(2, 1/3),list(5, diag(num_species))),
    sha_lt_params = 5,
    truth_params = list(10, list(3, 3), list(10, diag(num_species)))
)
}
\references{
Spence et. al. (2018). A general framework for combining ecosystem models. \emph{Fish and Fisheries}, 19(6):1031-1042.
}
