% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSample-class.R
\docType{class}
\name{EnsembleSample-class}
\alias{EnsembleSample-class}
\title{A class to hold the samples of the ensemble model from the Kalman filter}
\description{
\code{EnsembleSample} objects are generated using the \code{generate_sample} function.
}
\details{
The \code{mle} and \code{sample} slots are \eqn{T \times (M + 2)\times N_{sample}} arrays where
\eqn{T} is the time the ensemble model was run for, \eqn{M} is the number of simulators
and \eqn{N_{sample}} is the number of samples from the Kalman filter. For each time step,
the \eqn{t}th element of the array \eqn{S[t, ,]} is a matrix where each column is a
sample and the rows are the variables:
\deqn{\left( y^{(t)}, \eta^{(t)}, z_1^{(t)}, z_2^{(t)} \ldots, z_M^{(t)}   \right)}
where \eqn{y^{(t)}} is the ensemble model's prediction of the latent truth value at time \eqn{t},
\eqn{\eta^{(t)}} is the shared short-term discrepancy at time \eqn{t},
\eqn{z_i^{(t)}} is the individual short-term discrepancy of simulator \eqn{i} at time \eqn{t}.

Each sample in the \code{sample} slot is a sample from the Kalman filter, while each sample in
the \code{mle} slot is the MLE for the Kalman filter for every given sample of the ensemble model.
}
\section{Slots}{

\describe{
\item{\code{ensemble_fit}}{The \code{EnsembleFit} object of the fitted ensemble model.}

\item{\code{mle}}{The MLE point estimates for the Kalman filter.}

\item{\code{sample}}{Sample from the Kalman filter.}
}}

