% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSample-class.R
\name{EnsembleSample}
\alias{EnsembleSample}
\title{Constructor for the \code{EnsembleSample} class}
\usage{
EnsembleSample(ensemble_fit, mle, sample)
}
\arguments{
\item{ensemble_fit}{The \code{EnsembleFit} object containing the fitted ensemble model.}

\item{mle}{The MLE point estimate of the Kalman filter using the \code{ensemble_fit} object.}

\item{sample}{A sampling of the Kalman filter using the \code{ensemble_fit} object.}
}
\value{
An object of type \linkS4class{EnsembleSample}
}
\description{
A constructor for the \code{EnsembleSample} class. This is a container for the samples or
and MLE point estimates generated by the Kalman filter
}
