% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior_sampling.R
\name{prior_ensemble_model}
\alias{prior_ensemble_model}
\title{Generate samples of parameters from prior distribution}
\usage{
prior_ensemble_model(priors, M = 1, full_sample = TRUE, ...)
}
\arguments{
\item{priors}{An \code{EnsemblePrior} object specifying the prior distributions for the ensemble.}

\item{M}{A \code{numeric} that represents the number of simulators. The default is 1.}

\item{full_sample}{A \code{logical} that runs a full sampling of the prior density of the ensemble model if \code{TRUE}. If \code{FALSE}, returns the point estimate which maximises the prior density of the ensemble model.}

\item{...}{Additional arguments passed to the function \code{rstan::sampling} or  \code{rstan::optimizing}.}
}
\value{
A \code{list} containing two items named \code{samples} and \code{point_estimate}. If \code{full_sample==TRUE}, \code{samples} is a \code{stanfit} and \code{point_estimate} is a \code{NULL} object, else \code{samples} is a \code{NULL} and \code{point_estimate} is a \code{list} object.
}
\description{
Methods to generates samples of the parameters from the prior distribution of the ensemble model.
}
\examples{
\donttest{
num_species <- 4
priors <- EnsemblePrior(
 d = num_species,
 ind_st_params = IndSTPrior("lkj",  list(3, 2), 3, AR_params = c(1,1)),
 ind_lt_params = IndLTPrior(
   "beta",
   list(c(10,4,8, 7),c(2,3,1, 4)),
   list(matrix(5, num_species, num_species),
        matrix(0.5, num_species, num_species))
 ),
 sha_st_params = ShaSTPrior("inv_wishart",list(2, 1/3),list(5, diag(num_species))),
 sha_lt_params = 5,
 truth_params = TruthPrior(num_species, 10, list(3, 3), list(10, diag(num_species)))
)
prior_density <- prior_ensemble_model(priors, M = 3)
}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
\code{\linkS4class{EnsembleFit}}
}
