% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_ewe}
\alias{Sigma_ewe}
\title{Ecopath with EcoSim Sigma}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
Sigma_ewe
}
\description{
A 4x4 covariance matrix quantifying the parameter uncertainty of Ecopath with EcoSim
}
\keyword{datasets}
