% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_fs}
\alias{Sigma_fs}
\title{FishSUMS Sigma}
\format{
A 3x3 \code{matrix}.
}
\usage{
Sigma_fs
}
\description{
A 3x3 covariance matrix quantifying the parameter uncertainty of FishSUMS
}
\references{
Spence, M. A., Blanchard, J. L., Rossberg, A. G., Heath, M. R., Heymans, J. J., Mackinson, S., Serpetti, N., Speirs, D. C., Thorpe, R. B., and Blackwell, P. G. (2018). A general framework for combining ecosystem models. Fish and Fisheries, 19(6), 1031–1042.
}
\keyword{datasets}
