% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_lm}
\alias{Sigma_lm}
\title{LeMans Sigma}
\format{
A 4x4 \code{matrix}.
A 4x4 covariance matrix quantifying the parameter uncertainty of LeMans
}
\usage{
Sigma_lm
}
\description{
LeMans Sigma
}
\references{
Thorpe, R. B., Le Quesne, W. J. F., Luxford, F., Collie, J. S., and Jennings, S. (2015). Evaluation and management implications of uncertainty in a multispecies size-structured model of population and community responses to fishing. Methods in Ecology and Evolution, 6(1), 49–58.
}
\keyword{datasets}
