% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_miz}
\alias{Sigma_miz}
\title{mizer Sigma}
\format{
A 4x4 \code{matrix}.
A 4x4 covariance matrix quantifying the parameter uncertainty of mizer
}
\usage{
Sigma_miz
}
\description{
mizer Sigma
}
\references{
F. Scott, J.L. Blanchard and K.H. Andersen. mizer: an R package for
multispecies, trait-based and community size spectrum ecological
modelling. Methods in Ecology and Evolution 5(10) 1121-1125 (2014).
}
\keyword{datasets}
