% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_ewe}
\alias{SSB_ewe}
\title{Ecopath with EcoSim SSB}
\format{
A \code{data.frame} with 60 rows and 4 variables:
\describe{
\item{\code{N.pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
\item{\code{Sole}}{Spawning stock biomass of Sole, in log tonnes.}
}
}
\usage{
SSB_ewe
}
\description{
A dataset containing the predictions for spawning stock biomass of Norway Pout, Herring, Cod, and Sole in the North Sea between 1991-2050 under an MSY fishing scenario from EcoPath with EcoSim.
}
\references{
ICES (2016). Working Group on Multispecies Assessment Methods (WGSAM). Technical report, International Council for Exploration of the Seas.
}
\keyword{datasets}
