% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_fs}
\alias{SSB_fs}
\title{FishSUMS SSB}
\format{
A \code{data.frame} with 67 rows and 3 variables:
\describe{
\item{\code{N.Pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
}
}
\usage{
SSB_fs
}
\description{
A \verb{data frame} containing the predictions for spawning stock biomass of Norway Pout, Herring, and Cod
in the North Sea between 1984-2050 under an MSY fishing scenario from FishSUMS. Note that FishSUMS does not
produce outputs for Sole.
}
\references{
Speirs, D.C., Guirey, E.J., Gurney, W.S.C. and Heath, M.R. (2010).
A length structured partial ecosystem model for cod in the North Sea.
Fisheries Research 106, 474-494.
}
\keyword{datasets}
