% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_miz}
\alias{SSB_miz}
\title{mizer SSB}
\format{
A data frame with 67 rows and 4 variables:
\describe{
\item{\code{N.Pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
\item{\code{Sole}}{Spawning stock biomass of Sole, in log tonnes.}
}
}
\usage{
SSB_miz
}
\description{
A \code{data.frame} containing the predictions for spawning stock biomass of Norway Pout, Herring, Cod, and Sole
in the North Sea between 1984-2050 under an MSY fishing scenario from mizer.
}
\references{
F. Scott, J.L. Blanchard and K.H. Andersen. mizer: an R package for
multispecies, trait-based and community size spectrum ecological
modelling. Methods in Ecology and Evolution 5(10) 1121-1125 (2014).
}
\keyword{datasets}
