% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_obs}
\alias{SSB_obs}
\title{Stock assessment SSB}
\format{
A data frame with 34 rows and 4 variables:
\describe{
\item{\code{N.Pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
\item{\code{Sole}}{Spawning stock biomass of Sole, in log tonnes.}
}
}
\usage{
SSB_obs
}
\description{
A \code{data.frame} containing the single species stock assessment estimates of spawning stock
biomass of Norway Pout, Herring, Cod, and Sole in the North Sea between 1984-2017.
}
\references{
Herring Assessment Working Group for the Area South of 62 N (HAWG).Technical report,
ICES Scientific Reports. ACOM:07. 960 pp, ICES, Copenhagen.

Report of the Working Group on the Assessment of Demersal Stocks inthe North Sea and
Skagerrak. Technical report, ICES Scientific Reports. ACOM:22.pp, ICES, Copenhagen.
}
\keyword{datasets}
