% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_functions.R
\name{define_priors}
\alias{define_priors}
\title{Ensemble Discrepancy Priors}
\usage{
define_priors(
  ind_st_var_params,
  ind_st_cor_form,
  ind_st_cor_params,
  ind_lt_var_params,
  ind_lt_cor_form,
  ind_lt_cor_params,
  sha_st_var_exp,
  sha_st_cor_form,
  sha_st_cor_params,
  sha_lt_sd
)
}
\arguments{
\item{ind_st_var_params}{A \code{numeric} of length 2 giving the shape
and scale parameters of the inverse gamma prior on the variance of the individual
short-term discrepancies.}

\item{ind_st_cor_form}{A \code{character} describing the form of the correlation prior
used on the individual short-term discrepancies. Currently supported forms are
\code{'lkj'}, \code{'inv_wishart'}, or \code{'beta'}. See details below.}

\item{ind_st_cor_params}{A \code{list} giving the parameters for the parameters for the prior on the correlation matrix of
the individual short-term discrepancies. See details below.}

\item{ind_lt_var_params}{A \code{vector} of length 2 giving the shape
and scale parameters of the inverse gamma prior on the variance of the individual
long-term discrepancies.}

\item{ind_lt_cor_form}{A \code{character} describing the form of the correlation prior
used on the individual long-term discrepancies. Currently supported forms are
\code{'lkj'}, \code{'inv_wishart'}, or \code{'beta'}. See details below.}

\item{ind_lt_cor_params}{A \code{list} giving the parameters for the parameters for the prior on the correlation matrix of
the individual long-term discrepancies. See details below.}

\item{sha_st_var_exp}{The exponential parameter for the prior on the variance of
the shared short-term discrepancies.}

\item{sha_st_cor_form}{A \code{character} describing the form of the correlation prior
used on the shared short-term discrepancies. Currently supported forms are
\code{'lkj'}, \code{'inv_wishart'}, or \code{'beta'}. See details below.}

\item{sha_st_cor_params}{A \code{list} giving the parameters for the prior on the correlation matrix of
the shared short-term discrepancies. See details below.}

\item{sha_lt_sd}{A \code{numeric} of the standard deviation of the shared long-term discrepancy. Each
element corresponds to the standard deviation of each feature.}
}
\value{
A \code{list} encoding prior information on discrepancies.
}
\description{
Encodes prior information on discrepancies for use in the ensemble model.
See the supporting information for Spence et. al. (2018) for a worked example of this.
}
\details{
As in Spence et. al. (2018) , the discrepancy covariance matrices (individual and shared short-term discrepancies
as well as individual long-term discrepancies) are decomposed
into
\deqn{\Sigma = \sqrt{\mathrm{diag}(\pi_i)} \Lambda \sqrt{\mathrm{diag}(\pi)},}
where \eqn{\pi} is the vector of variances of each species, and \eqn{\Lambda}
is the correlation matrix. The variance terms \eqn{\pi}
are parameterised by inverse-gamma distributions, and passed through as \code{ind_st_var_params},
\code{ind_lt_var_params}, \code{sha_st_var_params} parameters.
}
\section{Correlation matrix priors}{

There are currently 3 supported prior distributions on correlation matrices which are chosen
via the \code{ind_st_cor_form}, \code{ind_lt_cor_form}, and \code{sha_st_cor_form} options.
These can take the value \code{'lkj'}, \code{'inv_wishart'}, or \code{'beta'} to
refer to the LKJ, inverse Wishart, or Beta distributions respectively. In each case,
the associated parameters should be passed through using the relevant \code{..._cor_params} variable.
\itemize{
\item LKJ - The parameter should be a single scalar value \eqn{\eta} giving the LKJ shape parameter as described in the
\href{https://mc-stan.org/docs/2_28/functions-reference/lkj-correlation.html}{Stan manual.} The
density of a correlation matrix is given by \deqn{f(\Sigma | \eta) \alpha det (\Sigma)^{\eta - 1}}
\item Inverse Wishart - A \code{list} containing a scalar value \eqn{\nu} (giving the degrees of
freedom) and a symmetric, positive definite matrix \eqn{\Sigma} (giving the scale
matrix). See the
\href{https://mc-stan.org/docs/2_28/functions-reference/inverse-wishart-distribution.html}{Stan manual}
for more information. The dimensions of \eqn{\Sigma} should be the same as the correlation matrix
it produces (i.e \eqn{N \times N} where \eqn{N} is the number of species)
\item Beta - A \code{list} containing two \eqn{N \times N} matrices (where \eqn{N} is the number of species), giving the prior success parameters \eqn{\alpha}
and prior failure parameters \eqn{\beta} respectively. To ensure positive-definiteness, the
correlations are rescaled from \eqn{[-1,1] \rightarrow [0,1]} via the function
\eqn{\frac{1}{\pi} \tan^{-1} \frac{\rho}{\sqrt{1-\rho^2} + 1/2}}. It is on these
rescaled parameters that the Beta distribution applies.
}
}

\examples{
#Basic usage of function for a model with 4 species.
N_species <- 4
priors <- define_priors(ind_st_var_params = list(25, 0.25),
                        ind_st_cor_form = "lkj", #Using an LKJ distribution for individual short-term discrepancies
                        ind_st_cor_params = 30, #The parameter is 30
                        ind_lt_var_params = list(rep(25,N_species),rep(0.25,N_species)),
                        ind_lt_cor_form = "beta", #For long term discrepancies we use a Beta distribution
                        ind_lt_cor_params = list(matrix(40,N_species, N_species), matrix(40, N_species, N_species)),

                        sha_st_var_exp = 3,
                        sha_st_cor_form = "lkj",
                        sha_st_cor_params = 30,
                        sha_lt_sd = rep(4,N_species))

}
\references{
Spence et. al. (2018). A general framework for combining ecosystem models. \emph{Fish and Fisheries}, 19(6):1031-1042.

Chandler RE. 2013 Exploiting strength, discounting weakness: combining information from multiple climate simulators. \emph{Phil Trans R Soc A} 371: 20120388
}
