% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EcoEnsemble-package.R
\docType{package}
\name{EcoEnsemble-package}
\alias{EcoEnsemble-package}
\alias{EcoEnsemble}
\title{Bayesian ensemble functions for combining ecosystem models}
\description{
The \code{EcoEnsemble} package implements the general ensemble framework for
ecosystem models laid out in Spence et al (2018) in R.
}
\details{
The ensemble model can be implemented in three main stages:
\enumerate{
\item Eliciting priors on discrepancy terms. This is done by using the
\code{elicit_priors} function.
\item Fitting the ensemble model. Using \code{fit_ensemble_model} with simulator
outputs, observations, and priors, the ensemble model can be fit, obtaining
either the point estimate which maximises the posterior, or running MCMC to
generate a sample from the posterior of the ensemble model.
\item Sampling from the Kalman filter. Using \code{generate_sample} with
the fitted ensemble object, the discrepancy terms and the ensemble's best
guess of the truth can be generated. Similarly to \code{fit_ensemble_model}, this
can either be a point estimate (the MLE) or a full sample of the Kalman filter.
}
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
Spence et. al. (2018). A general framework for combining ecosystem models. https://onlinelibrary.wiley.com/doi/abs/10.1111/faf.12310
}
