% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFit-class.R
\docType{class}
\name{EnsembleFit-class}
\alias{EnsembleFit-class}
\title{A class to hold the samples or point estimates from the ensemble model.}
\description{
An \code{EnsembleFit} object is returned by the \code{fit_ensemble_model} function.
The object contains a slot for the \code{EnsembleData} object originally used to
fit the ensemble model. The \code{samples} slot contains the samples from the MCMC
if a full sampling was completed, otherwise the \code{point_estimate} slot contains
information about a point estimate.
}
\section{Slots}{

\describe{
\item{\code{ensemble_data}}{The \code{EnsembleData} object encapsulating the data initially passed
through to Stan when fitting the ensemble model.}

\item{\code{samples}}{The \code{stanfit} object representing the samples drawn from the Stan
model. This is \code{NULL} if only a point estimate is generated.}

\item{\code{point_estimate}}{The \code{list} output of \code{rstan::optimizing} if a full sampling is
not completed. Otherwise, this is \code{NULL}.}
}}

