% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleFit-class.R
\name{EnsembleFit}
\alias{EnsembleFit}
\title{Constructor for the \code{EnsembleFit} class}
\usage{
EnsembleFit(ensemble_data, samples = NULL, point_estimate = NULL)
}
\arguments{
\item{ensemble_data}{The \code{EnsembleData} object encapsulating the data initially passed
through to Stan when fitting the ensemble model.}

\item{samples}{The \code{stanfit} object representing the samples drawn from the Stan model.}

\item{point_estimate}{The \code{list} output of \code{rstan::optimizing} if a full sampling is
not completed.}
}
\value{
An object of type \linkS4class{EnsembleFit}
}
\description{
A constructor for the \code{EnsembleFit} class. This is a container for the samples or
point estimates generated by the fitted Stan models.
}
