% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ensemble.R
\name{fit_ensemble_model}
\alias{fit_ensemble_model}
\title{Fits the ensemble model}
\usage{
fit_ensemble_model(observations, simulators, priors, full_sample = TRUE, ...)
}
\arguments{
\item{observations}{A \code{list} of length 2 containing observations and a covariance matrix. The first element is a \code{data.frame} or \code{matrix} with each column giving observations of each output of interest and each row a time. Rows should be named with the times and columns should be named the variables. The second element is the covariance matrix of the observations.}

\item{simulators}{A \code{list} with length equal to the number of simulators. For each simulator, there is a \code{list} of 2 objects containing the simulator output and covariance matrix. The first element is a \code{data.frame} or \code{matrix} with each column giving a simulator outputs of interest and each row a time. Rows should be named with the times and columns should be named the variables. The second element is the covariance matrix of the simulator outputs.}

\item{priors}{A \code{list} of data specifying the prior distributions for the ensemble.
See the \code{define_priors} function for the required form.}

\item{full_sample}{A \code{logical} that runs a full sampling of the ensemble model if \code{TRUE} and optimises to the MLE if \code{FALSE}. The default is TRUE.}

\item{...}{Additional parameters passed to the function \code{rstan::sampling}.}
}
\value{
An \code{EnsembleFit} object containing the data passed through to  and the fitted \code{stanfit} object.
}
\description{
\code{fit_ensemble_model} runs an MCMC of the ensemble model. This process can take a long time depending on the size of the datasets.
}
\examples{
N_species <- 4
priors <- define_priors(ind_st_var_params = list(25, 0.25),
                        ind_st_cor_form = "lkj",
                        ind_st_cor_params = 30,
                        ind_lt_var_params = list(rep(25,N_species),rep(0.25,N_species)),
                        ind_lt_cor_form = "beta",
                        ind_lt_cor_params = list(matrix(40,N_species, N_species), matrix(40, N_species, N_species)),

                        sha_st_var_exp = 3,
                        sha_st_cor_form = "lkj",
                        sha_st_cor_params = 30,
                        sha_lt_sd = rep(4,N_species))
fit <- fit_ensemble_model(observations = list(SSB_obs, Sigma_obs),
                         simulators = list(list(SSB_ewe, Sigma_ewe, "EwE"),
                                     list(SSB_fs,  Sigma_fs, "FishSUMS"),
                                     list(SSB_lm,  Sigma_lm, "LeMans"),
                                     list(SSB_miz, Sigma_miz, "Mizer")),
                         priors = priors,
                         full_sample = FALSE) #Only optimise in this case
Run a full sample
fit1 <- fit_ensemble_model(observations = list(SSB_obs, Sigma_obs),
                         simulators = list(list(SSB_ewe, Sigma_ewe, "EwE"),
                                     list(SSB_fs,  Sigma_fs, "FishSUMS"),
                                     list(SSB_lm,  Sigma_lm, "LeMans"),
                                     list(SSB_miz, Sigma_miz, "Mizer")),
                         priors = priors,
                         control=list(adapt_delta = 0.99)) # Additional Stan parameters.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
\seealso{
\code{\linkS4class{EnsembleFit}}, \code{\link{EnsembleSample}}
}
