% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ensemble.R
\name{get_mcmc_ensemble_model}
\alias{get_mcmc_ensemble_model}
\title{Return the compiled ensemble model STAN object.}
\usage{
get_mcmc_ensemble_model()
}
\value{
The \code{stanmodel} object encoding the ensemble model.
}
\description{
Gets the unfit, compiled STAN object encoding the ensemble model. This allows for
manual fitting of the ensemble model directly using.
}
\examples{
#A minimal example fitting manually
mod <- get_mcmc_ensemble_model()
stan_data <- get_stan_ensemble_data(observations = list(SSB_obs, Sigma_obs),
                                   simulators = list(list(SSB_ewe, Sigma_ewe, "EwE"),
                                                     list(SSB_fs,  Sigma_fs),
                                                     list(SSB_lm,  Sigma_lm),
                                                     list(SSB_miz, Sigma_miz)),
                                   priors = priors)
out <- rstan::sampling(mod, stan_data, chains=(parallel::detectCores()-1))
}
