% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_ewe}
\alias{Sigma_ewe}
\title{Ecopath with EcoSim Sigma}
\format{
A 4x4 \code{matrix}.
}
\usage{
Sigma_ewe
}
\description{
A 4x4 covariance matrix quantifying the parameter uncertainty of Ecopath with EcoSim
}
\references{
Mackinson, S., Platts, M., Garcia, C., and Lynam, C. (2018). Evaluating the fishery and ecological consequences of the proposed North Sea multi-annual plan. PLOS ONE, 13(1), 1–23.
}
\keyword{datasets}
