% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_ewe}
\alias{Sigma_ewe}
\title{Ecopath with EcoSim Sigma}
\format{
A 4x4 \code{matrix}.
}
\usage{
Sigma_ewe
}
\description{
A 4x4 covariance matrix quantifying the parameter uncertainty of Ecopath with EcoSim
}
\references{
C Walters, V Christensen and D Pauly (1997) Structuring dynamic
models of exploited ecosystems from trophic mass-balance assessments
Rev. Fish Biol. Fish., 7, pp. 139-172
}
\keyword{datasets}
