% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_fs}
\alias{Sigma_fs}
\title{FishSUMS Sigma}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3 rows and 3 columns.
}
\usage{
Sigma_fs
}
\description{
A 3x3 covariance matrix quantifying the parameter uncertainty of FishSUMS
}
\keyword{datasets}
