% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_miz}
\alias{Sigma_miz}
\title{mizer Sigma}
\format{
A 4x4 \code{matrix}.
A 4x4 covariance matrix quantifying the parameter uncertainty of mizer
}
\usage{
Sigma_miz
}
\description{
mizer Sigma
}
\references{
Spence, M. A., Blackwell, P. G., and Blanchard, J. L. (2016). Parameter uncertainty of a dynamic multispecies size spectrum model. Canadian Journal of Fisheries and Aquatic Sciences, 73(4), 589–59
}
\keyword{datasets}
