% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sigma_obs}
\alias{Sigma_obs}
\title{Stock assessment Sigma}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 4 rows and 4 columns.
}
\usage{
Sigma_obs
}
\description{
A 4x4 covariance matrix quantifying the covariances of the stock assessment estimates of biomass.
}
\keyword{datasets}
