% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_ewe}
\alias{SSB_ewe}
\title{Ecopath with EcoSim SSB}
\format{
A \code{data.frame} with 60 rows and 4 variables:
\describe{
\item{\code{N.Pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
\item{\code{Sole}}{Spawning stock biomass of Sole, in log tonnes.}
}
}
\usage{
SSB_ewe
}
\description{
A dataset containing the predictions for spawning stock biomass of Norway Pout, Herring, Cod,
and Sole in the North Sea between 1991-2050 under an MSY fishing scenario from EcoPath with EcoSim.
}
\references{
C Walters, V Christensen and D Pauly (1997) Structuring dynamic
models of exploited ecosystems from trophic mass-balance assessments
Rev. Fish Biol. Fish., 7, pp. 139-172
}
\keyword{datasets}
