% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SSB_lm}
\alias{SSB_lm}
\title{LeMans SSB}
\format{
A \code{data.frame} with 65 rows and 4 variables:
\describe{
\item{\code{N.Pout}}{Spawning stock biomass of Norway Pout, in log tonnes.}
\item{\code{Herring}}{Spawning stock biomass of Herring, in log tonnes.}
\item{\code{Cod}}{Spawning stock biomass of Cod, in log tonnes.}
\item{\code{Sole}}{Spawning stock biomass of Sole, in log tonnes.}
}
}
\usage{
SSB_lm
}
\description{
A \code{data.frame} containing the predictions for spawning stock biomass of Norway Pout, Herring, Cod, and Sole
in the North Sea between 1986-2050 under an MSY fishing scenario from the LeMaRns package.
}
\references{
Michael A. Spence, Hayley J. Bannister, Johnathan E. Ball, Paul J. Dolder
and Robert B. Thorpe (2019). LeMaRns: Length-Based Multispecies Analysis
by Numerical Simulation. R package version 0.1.2.
\href{https://CRAN.R-project.org/package=LeMaRns}{CRAN}
}
\keyword{datasets}
